notification_service/notification_service/requirements.txt
django>=4.2
djangorestframework>=3.14
psycopg2-binary>=2.9
django-filter>=23.0
pika>=1.3